requirements.txt
flask
langchain
langchain-google-genai
google-generativeai
chromadb
docx2txt
python-docx
gradio
python-dotenv